# Improve speed/accuracy of FasterRCNN by introducing a score threshold on RPN