--- a/test/test_onnx.py
+++ b/test/test_onnx.py
@@ -197,7 +197,7 @@ class ONNXExporterTester(unittest.TestCase):
         rpn_pre_nms_top_n = dict(training=2000, testing=1000)
         rpn_post_nms_top_n = dict(training=2000, testing=1000)
         rpn_nms_thresh = 0.7
-        rpn_score_thresh = 0.05
+        rpn_score_thresh = 0.0
 
         rpn = RegionProposalNetwork(
             rpn_anchor_generator, rpn_head,
@@ -368,7 +368,8 @@ class ONNXExporterTester(unittest.TestCase):
     def test_faster_rcnn(self):
         images, test_images = self.get_test_images()
         dummy_image = [torch.ones(3, 100, 100) * 0.3]
-        model = models.detection.faster_rcnn.fasterrcnn_resnet50_fpn(pretrained=True, min_size=200, max_size=300)
+        model = models.detection.faster_rcnn.fasterrcnn_resnet50_fpn(pretrained=True, min_size=200, max_size=300,
+                                                                     rpn_score_thresh=0.0)
         model.eval()
         model(images)
         # Test exported model on images of different size, or dummy input
@@ -419,7 +420,8 @@ class ONNXExporterTester(unittest.TestCase):
     def test_mask_rcnn(self):
         images, test_images = self.get_test_images()
         dummy_image = [torch.ones(3, 100, 100) * 0.3]
-        model = models.detection.mask_rcnn.maskrcnn_resnet50_fpn(pretrained=True, min_size=200, max_size=300)
+        model = models.detection.mask_rcnn.maskrcnn_resnet50_fpn(pretrained=True, min_size=200, max_size=300,
+                                                                 rpn_score_thresh=0.0)
         model.eval()
         model(images)
         # Test exported model on images of different size, or dummy input
@@ -469,7 +471,8 @@ class ONNXExporterTester(unittest.TestCase):
     def test_keypoint_rcnn(self):
         images, test_images = self.get_test_images()
         dummy_images = [torch.ones(3, 100, 100) * 0.3]
-        model = models.detection.keypoint_rcnn.keypointrcnn_resnet50_fpn(pretrained=True, min_size=200, max_size=300)
+        model = models.detection.keypoint_rcnn.keypointrcnn_resnet50_fpn(pretrained=True, min_size=200, max_size=300,
+                                                                         rpn_score_thresh=0.0)
         model.eval()
         model(images)
         self.run_model(model, [(images,), (test_images,), (dummy_images,)],