@datumbox 2021-01-07
@t-vi I discussed with @fmassa about adding the specific threshold in the FasterRCNN implementation as it seems to benefit both the speed and the accuracy metrics and your name came up. Any thoughts on this?