@masahi 2021-01-07
well not really an opinion, but nothing changes by default and the only extra costs are one additional parameter (of which there are already many) and one call to sigmoid and where, both negligible. So I'd say why not.