@datumbox 2021-01-14
The failing tests on ONNX is not related to this PR. The master is broken (see #3251).
The failing clang_format test is again unrelated and it's due to a version change on the formatter that is being investigated.